style: ruff formatting

+++ utils/genome_generation.py
@@ -1,3 +1,5 @@
+# ruff: noqa: S321
+
 import argparse
 import csv
 import ftplib
@@ -11,7 +13,6 @@ import subprocess
 import urllib.request
 from functools import cache
 from http.client import HTTPResponse
-from typing import Union
 
 _ncbi_url = "https://api.ncbi.nlm.nih.gov/datasets/v2alpha"
 _ensembl_url = "ftp.ensembl.org"
@@ -19,14 +20,19 @@ _ensembl_url = "ftp.ensembl.org"
 _ucsc_url = "https://api.genome.ucsc.edu"
 _ref_flat_url = "https://hgdownload.soe.ucsc.edu/goldenPath/{final_genome}/database/refFlat.txt.gz"  # fmt: skip
 
+MUS_MUSCULUS: int = 10090
+HOMO_SAPIENS: int = 9606
+MACACA_MULATTA: int = 9544
+MIN_ENSEMBL_RELEASE: int = 19
+
 
 class Utilities:
-    _latest_release: Union[int, None] = None
+    _latest_release: int | None = None
+    _species_name: str | None = None
 
     @staticmethod
     def get_latest_release() -> int:
-        """
-        This function will identify the latest `release-###` from ensembl
+        """Identify the latest `release-###` from ensembl.
 
         Returns:
             int: An integer matching the release number. For example, if the latest release is `release-112`, this function returns `112`
@@ -42,41 +48,37 @@ class Utilities:
         for release in pub_root.nlst("/pub"):
             if release.startswith("/pub/release-") and not release.endswith(".txt"):
                 release_number: int = int(release.split("-")[-1])
-                if release_number > latest_release:
-                    latest_release = release_number
+                latest_release = max(latest_release, release_number)
 
         Utilities._latest_release = latest_release
         return latest_release
 
     @staticmethod
     @cache
-    def is_validate_release_number(release_number: Union[str, int]) -> bool:
-        """
-        This function will make sure `release_number` exists in the ensembl_ftp_url
+    def is_validate_release_number(release_number: str | int) -> bool:
+        """Validate `release_number` exists in the ensembl_ftp_url.
 
         Args:
-            release_number (str): The release number to check. If a string is provided, it should be in the format of (for example) `release-112`. If an integer is provided, it should only be the number (for example) `112`
+            release_number (str): The release number to check.
+                If a string is provided, it should be in the format of (for example) `release-112`.
+                If an integer is provided, it should only be the number (for example) `112`
 
         Returns:
             bool: True if the release number is valid, False otherwise
         """
         try:
             if isinstance(release_number, str):
-                if release_number.startswith("release-"):
-                    release_number = int(release_number.split("-")[-1])
-                else:
-                    release_number = int(release_number)
+                release_number = int(release_number.split("-")[-1]) if release_number.startswith("release-") else int(release_number)
         except ValueError:  # Unable to convert to an integer
             return False
 
         latest_release = Utilities.get_latest_release()
-        return 19 <= release_number <= latest_release  # 19 is the earliest release number
+        return MIN_ENSEMBL_RELEASE <= release_number <= latest_release  # 19 is the earliest release number
 
     @staticmethod
     @cache
     def is_valid_taxon(taxon_id: int) -> bool:
-        """
-        This function will make sure the species provided is correct by validating it against NCBI's Taxonomy database
+        """Validate the provided species is correct by checking it against NCBI's Taxonomy database.
 
         Args:
             taxon_id (int): The NCBI Taxonomy ID
@@ -85,12 +87,12 @@ class Utilities:
             bool: True if the species is valid, False otherwise
         """
         invalid_taxon_error = f"The taxonomy name you specified ({taxon_id}) is not a recognized NCBI Taxonomy name."
-        response: HTTPResponse = urllib.request.urlopen(f"{_ncbi_url}/taxonomy/taxon/{taxon_id}/name_report")
+        response: HTTPResponse = urllib.request.urlopen(f"{_ncbi_url}/taxonomy/taxon/{taxon_id}/name_report")  # noqa: S310
         as_json: dict = json.loads(response.read())
 
         is_valid = True
         for response_list in as_json["reports"]:
-            if "errors" in response_list.keys():
+            if "errors" in response_list:
                 error_keys = response_list["errors"][0].keys()
                 error_values = response_list["errors"][0].values()
 
@@ -103,12 +105,12 @@ class Utilities:
     @cache
     def get_species_from_taxon(
         taxon_id: int,
+        *,
         lowercase: bool = True,
         replace_spaces: bool = True,
         replace_with: str = "_",
     ) -> str:
-        """
-        This function will return the species name from the provided taxon ID
+        """Get the species name from the provided taxon ID.
 
         It assumes the taxon ID is valid, as defined by `_is_valid_taxon`
 
@@ -121,16 +123,34 @@ class Utilities:
         Returns:
             str: The species name that corresponds to the ensembl species. For example: 9606 -> homo_sapiens
         """
-        response: HTTPResponse = urllib.request.urlopen(f"{_ncbi_url}/taxonomy/taxon/{taxon_id}/name_report")
-        as_json = json.loads(response.read())
-        species = as_json["reports"][0]["taxonomy"]["current_scientific_name"]["name"]
+        # Hardcode common, known species IDs
+        if taxon_id == MUS_MUSCULUS:
+            species = "Mus musculus"
+        elif taxon_id == HOMO_SAPIENS:
+            species = "Homo sapiens"
+        elif taxon_id == MACACA_MULATTA:
+            species = "Macaca mulatta"
 
-        if lowercase:
-            species = species.lower()
-        if replace_spaces:
-            species = species.replace(" ", replace_with)
+        if species in {"Mus musculus", "Homo sapiens", "Macaca mulatta"}:
+            if lowercase:
+                species = species.lower()
+            if replace_spaces:
+                species = species.replace(" ", replace_with)
+            return species
 
-        return species
+        if Utilities._species_name is None:
+            response: HTTPResponse = urllib.request.urlopen(f"{_ncbi_url}/taxonomy/taxon/{taxon_id}/name_report")  # noqa: S310
+            as_json = json.loads(response.read())
+            species = as_json["reports"][0]["taxonomy"]["current_scientific_name"]["name"]
+
+            if lowercase:
+                species = species.lower()
+            if replace_spaces:
+                species = species.replace(" ", replace_with)
+
+            Utilities._species_name = species
+
+        return Utilities._species_name
 
 
 class NCBI:
@@ -138,6 +158,7 @@ class NCBI:
         self,
         taxon_id: int,
         release_number: str,
+        *,
         show_progress: bool,
     ) -> None:
         self._taxon_id: int = taxon_id
@@ -149,9 +170,14 @@ class NCBI:
         if not self._release_number.startswith(("latest", "release-")):
             raise ValueError("""release_number must start with 'latest' or 'release-'.\nExamples: "latest", "release-112""")  # fmt: skip
         if not Utilities.is_valid_taxon(self._taxon_id):
-            raise ValueError(f"The provided taxon_id ({self._taxon_id}) is not valid.\nPlease validate your taxon ID at https://www.ncbi.nlm.nih.gov/Taxonomy")  # fmt: skip
+            raise ValueError(
+                f"The provided taxon_id ({self._taxon_id}) is not valid.\nPlease validate your taxon ID at https://www.ncbi.nlm.nih.gov/Taxonomy"
+            )
         if self._release_number != "latest" and not Utilities.is_validate_release_number(self._release_number):
-            raise ValueError(f"The provided release number ({self._release_number}) is not valid. It should be in the format of `release-112` or `112`.\nPlease validate your release number at https://ftp.ensembl.org/pub")  # fmt: skip
+            raise ValueError(
+                f"The provided release number ({self._release_number}) is not valid. It should be in the format of `release-112` or `112`.\n"
+                f"Please validate your release number at https://ftp.ensembl.org/pub"
+            )
 
         # Inputs are valid, we can now download the fasta primary assembly
         if self._release_number == "latest":
@@ -161,19 +187,21 @@ class NCBI:
         self._ftp.login()
 
     def download_fasta_file(self, save_directory: str) -> None:
-        """
-        This function will download the reference genome from the UCSC Genome Browser and save it as a FASTA file.
+        """Download the reference genome from the UCSC Genome Browser and save it as a FASTA file.
 
         # The full input path of the genome fasta file
         GENOME_FASTA_FILE: "genome/Mus_musculus.GRCm39.dna.primary_assembly.fa"
         GTF_FILE: "genome/Mus_musculus.GRCm39.111.gtf"
         """
+        final_output_filepath = os.path.join(save_directory, self._species_name, f"{self._species_name}_{self._release_number}_primary_assembly.fa")
+        if os.path.exists(final_output_filepath):
+            return
+
         save_directory = os.path.join(save_directory, self._species_name)
         if not os.path.exists(save_directory):
             os.makedirs(save_directory, exist_ok=True)
 
         print(f"fasta file save directory: {save_directory}")
-
         fasta_root = f"/pub/{self._release_number}/fasta/{self._species_name}/dna"
         primary_assembly_suffix = ".dna.primary_assembly.fa.gz"
         primary_assembly_path = ""
@@ -190,7 +218,7 @@ class NCBI:
         # If the primary assembly path is not empty, we can download it
         if primary_assembly_path != "":
             primary_assembly_filename = primary_assembly_path.split("/")[-1]
-            with open(f"{save_directory}/{primary_assembly_filename}", "wb") as fasta_file:  # fmt: skip
+            with open(os.path.join(save_directory, primary_assembly_filename), "wb") as fasta_file:
                 self._ftp.retrbinary(f"RETR {primary_assembly_path}", fasta_file.write)
         # Otherwise we must download all *.dna.chromosome.*.fa.gz files and concatenate them
         else:
@@ -203,7 +231,7 @@ class NCBI:
                     primary_assembly_filename = filename.split("/")[-1].replace(".1.fa.gz", ".fa.gz")  # fmt: skip
 
             # TODO: add a progress bar to this download
-            with open(f"{save_directory}/{primary_assembly_filename}", "wb") as fasta_file:
+            with open(os.path.join(save_directory, primary_assembly_filename), "wb") as fasta_file:
                 for remote_chromosome in chromosome_files:
                     chromosome_filename = remote_chromosome.split("/")[-1]
                     local_chromosome = f"{save_directory}/{chromosome_filename}"
@@ -229,16 +257,18 @@ class NCBI:
 
         # Remove the gzipped file then convert file to lowercase name
         os.remove(f"{save_directory}/{primary_assembly_filename}")
-        shutil.move(
-            os.path.join(save_directory, txt_filename),
-            os.path.join(save_directory, f"{self._species_name}_{self._release_number}_primary_assembly.fa"),
-        )
+        shutil.move(os.path.join(save_directory, txt_filename), final_output_filepath)
 
     def download_gtf_file(self, save_directory: str) -> None:
         """
         # The full input path of the GTF file
         GTF_FILE: "genome/Mus_musculus.GRCm39.111.gtf"
         """
+        final_output_filepath = os.path.join(save_directory, self._species_name, f"{self._species_name}_{self._release_number}.gtf")
+        if os.path.exists(final_output_filepath):
+            print(f"GTF file exists, not creating! Checked path: '{final_output_filepath}'")
+            return
+
         save_directory = os.path.join(save_directory, self._species_name)
         if not os.path.exists(save_directory):
             os.makedirs(save_directory, exist_ok=True)
@@ -267,7 +297,7 @@ class NCBI:
         os.remove(f"{save_directory}/{gtf_filename}")
         shutil.move(
             os.path.join(save_directory, txt_filename),
-            os.path.join(save_directory, f"{self._species_name}_{self._release_number}.gtf"),
+            final_output_filepath,
         )
 
 
@@ -284,6 +314,11 @@ def ref_flat_file_creation(taxon_id: int, save_directory: str) -> None:
     save_directory = os.path.join(save_directory, species_name)
     if not os.path.exists(save_directory):
         os.makedirs(save_directory, exist_ok=True)
+
+    final_output_filepath = os.path.join(save_directory, f"{species_name}_ref_flat.txt")
+    if os.path.exists(final_output_filepath):
+        print(f"ref flat file exists, not creating! Checked path: '{final_output_filepath}'")
+        return
 
     print(f"ref flat file save directory: {save_directory}")
 
@@ -320,7 +355,7 @@ def ref_flat_file_creation(taxon_id: int, save_directory: str) -> None:
         with open(txt_file, "wb") as o_stream:
             o_stream.write(decompressed.read())
 
-    shutil.move(txt_file, os.path.join(save_directory, f"{species_name}_ref_flat.txt"))
+    shutil.move(txt_file, final_output_filepath)
 
 
 def rRNA_interval_list_creation(taxon_id: int, save_directory: str) -> None:
@@ -339,11 +374,15 @@ def rRNA_interval_list_creation(taxon_id: int, save_directory: str) -> None:
     Raises:
         FileNotFoundError: «The primary assembly file could not be found» if the primary assembly file is not found
     """
-
     species_name = Utilities.get_species_from_taxon(taxon_id)
     save_directory = os.path.join(save_directory, species_name)
     if not os.path.exists(save_directory):
         os.makedirs(save_directory, exist_ok=True)
+
+    final_output_filepath = os.path.join(save_directory, f"{species_name}_rrna.interval_list")
+    if os.path.exists(final_output_filepath):
+        print(f"rRNA interval list exists, not creating! Checked path: '{final_output_filepath}'")
+        return
 
     print(f"rRNA interval list save directory: {save_directory}")
 
@@ -369,9 +408,10 @@ def rRNA_interval_list_creation(taxon_id: int, save_directory: str) -> None:
         ["samtools", "faidx", primary_assembly_fa, primary_assembly_fai],
         stdout=subprocess.PIPE,
         stderr=subprocess.PIPE,
+        check=False,
     )
 
-    primary_assembly_fai_in = open(primary_assembly_fai, "r")
+    primary_assembly_fai_in = open(primary_assembly_fai)
     genome_sizes_out = open(genome_sizes, "w")
     for line in primary_assembly_fai_in:
         fields = line.split()
@@ -379,8 +419,8 @@ def rRNA_interval_list_creation(taxon_id: int, save_directory: str) -> None:
     primary_assembly_fai_in.close()
     genome_sizes_out.close()
 
-    genes_in = open(genes, "r").readlines()
-    genome_sizes_in = open(genome_sizes, "r").readlines()
+    genes_in = open(genes).readlines()
+    genome_sizes_in = open(genome_sizes).readlines()
     genome_sizes_in = [line.strip() for line in genome_sizes_in]
     rRNA_interval_list_out = open(rRNA_interval_list, "w")
     for line in genome_sizes_in:
@@ -393,11 +433,11 @@ def rRNA_interval_list_creation(taxon_id: int, save_directory: str) -> None:
                 gene_id = fields[-1].split('"')[1]
                 rRNA_interval_list_out.write(f"{fields[0]}\t{fields[3]}\t{fields[4]}\t{fields[6]}\t{gene_id}\n")
 
-    shutil.move(rRNA_interval_list, os.path.join(save_directory, f"{species_name}_rrna.interval_list"))
+    shutil.move(rRNA_interval_list, final_output_filepath)
 
     # Sort the interval list
     # TODO: Is this needed? As of now, I don't think so
-    # subprocess.run(["sort", "-k1V", "-k2n", "-k3n", "-o", rRNA_interval_list, rRNA_interval_list])  # fmt: skip
+    # subprocess.run(["sort", "-k1V", "-k2n", "-k3n", "-o", rRNA_interval_list, rRNA_interval_list])
 
 
 def bed_file_creation(taxon_id: int, save_directory: str) -> None:
@@ -415,12 +455,16 @@ def bed_file_creation(taxon_id: int, save_directory: str) -> None:
     if not os.path.exists(save_directory):
         os.makedirs(save_directory, exist_ok=True)
 
+    if os.path.exists(bed_output_file):
+        print(f"bed file exists, not creating! Checked path: '{bed_output_file}'")
+        return
+
     print(f"bed file save directory: {save_directory}")
 
     ref_flat_file = os.path.join(save_directory, f"{species_name}_ref_flat.txt")
     assert os.path.exists(ref_flat_file), f"The refFlat file could not be found\nSearching for: '{ref_flat_file}'"
 
-    reader = csv.reader(open(ref_flat_file, "r"), delimiter="\t", quotechar=None)
+    reader = csv.reader(open(ref_flat_file), delimiter="\t", quotechar=None)
     writer = csv.writer(open(bed_output_file, "w"), delimiter="\t", quotechar=None)
     writer.writerow(['track name="refflat"'])
 
@@ -437,7 +481,7 @@ def bed_file_creation(taxon_id: int, save_directory: str) -> None:
                 row[7],
                 0,
                 row[8],
-                ",".join([str(int(x) - int(y)) for x, y in zip(row[10].split(","), row[9].split(",")) if x and y]),
+                ",".join([str(int(x) - int(y)) for x, y in zip(row[10].split(","), row[9].split(","), strict=False) if x and y]),
                 ",".join([str(int(x) - int(row[4])) for x in row[9].split(",") if x]),
             )
         )
